Move CoverArtSetter to its own file
- it helps to reduce the number of imports per file
