+++ picard/ui/coverartbox/__init__.py
@@ -35,8 +35,6 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
 
-from contextlib import ExitStack
-from enum import IntEnum
 from functools import partial
 import os
 import re
@@ -49,108 +47,26 @@ from PyQt6 import (
 )
 
 from picard import log
-from picard.album import Album
-from picard.cluster import Cluster
 from picard.config import get_config
 from picard.coverart.image import (
     CoverArtImage,
     CoverArtImageError,
 )
-from picard.file import File
 from picard.i18n import gettext as _
-from picard.item import FileListItem
-from picard.track import Track
 from picard.util import (
     imageinfo,
     normpath,
 )
 from picard.util.lrucache import LRUCache
 
+from .coverartsetter import (
+    CoverArtSetter,
+    CoverArtSetterMode,
+)
 from .coverartthumbnail import CoverArtThumbnail
 from .imageurldialog import ImageURLDialog
 
 from picard.ui.util import FileDialog
-
-
-class CoverArtSetterMode(IntEnum):
-    APPEND = 0
-    REPLACE = 1
-
-
-class CoverArtSetter:
-
-    def __init__(self, mode, coverartimage, source_obj):
-        self.mode = mode
-        self.coverartimage = coverartimage
-        self.source_obj = source_obj
-
-        if isinstance(self.source_obj, Album):
-            self.set_coverart = self.set_coverart_album
-        elif isinstance(self.source_obj, FileListItem):
-            self.set_coverart = self.set_coverart_filelist
-        elif isinstance(self.source_obj, File):
-            self.set_coverart = self.set_coverart_file
-
-    def set_coverart(self):
-        return False
-
-    def set_image(self, obj):
-        if self.mode == CoverArtSetterMode.REPLACE:
-            obj.metadata.images.strip_front_images()
-
-        obj.metadata.images.append(self.coverartimage)
-        obj.metadata_images_changed.emit()
-
-    def set_coverart_album(self):
-        album = self.source_obj
-        with ExitStack() as stack:
-            stack.enter_context(album.suspend_metadata_images_update)
-            self.set_image(album)
-            for track in album.tracks:
-                stack.enter_context(track.suspend_metadata_images_update)
-                self.set_image(track)
-            for file in album.iterfiles():
-                self.set_image(file)
-                file.update(signal=False)
-        album.update(update_tracks=False)
-        return True
-
-    @staticmethod
-    def iter_file_parents(file):
-        parent = file.parent_item
-        if parent:
-            yield parent
-            if isinstance(parent, Track) and parent.album:
-                yield parent.album
-            elif isinstance(parent, Cluster) and parent.related_album:
-                yield parent.related_album
-
-    def set_coverart_filelist(self):
-        filelist = self.source_obj
-        parents = set()
-        with ExitStack() as stack:
-            stack.enter_context(filelist.suspend_metadata_images_update)
-            self.set_image(filelist)
-            for file in filelist.iterfiles():
-                for parent in self.iter_file_parents(file):
-                    stack.enter_context(parent.suspend_metadata_images_update)
-                    parents.add(parent)
-                self.set_image(file)
-                file.update(signal=False)
-            for parent in parents:
-                self.set_image(parent)
-                if isinstance(parent, Album):
-                    parent.update(update_tracks=False)
-                else:
-                    parent.update()
-        filelist.update()
-        return True
-
-    def set_coverart_file(self):
-        file = self.source_obj
-        self.set_image(file)
-        file.update()
-        return True
 
 
 HTML_IMG_SRC_REGEX = re.compile(r'<img .*?src="(.*?)"', re.UNICODE)
